pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "netfs"
version = "0.1.0"
description = "Remote filesystem and exec bridge over reverse SSH"
requires-python = ">=3.5"
dependencies = []

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]
fuse = [
  "fusepy>=3.0.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
